Add Resource.new(url: <url>) syntax back to yt
