+++ lib/yt.rb
@@ -13,7 +13,6 @@ require 'yt/models/video_group'
 require 'yt/models/comment_thread'
 require 'yt/models/ownership'
 require 'yt/models/advertising_options_set'
-require 'yt/models/url'
 
 # An object-oriented Ruby client for YouTube.
 # Helps creating applications that need to interact with YouTube objects.
