+++ spec/models/resource_spec.rb
@@ -4,6 +4,13 @@ require 'yt/models/resource'
 describe Yt::Resource do
   subject(:resource) { Yt::Resource.new attrs }
 
+  context 'given a resource initialized with a URL (containing an ID)' do
+    let(:attrs) { {url: 'youtu.be/9bZkp7q19f0'} }
+
+    it { expect(resource.id).to eq '9bZkp7q19f0' }
+    it { expect(resource.kind).to eq 'video' }
+  end
+
   describe '#public?' do
     context 'given fetching a status returns privacyStatus "public"' do
       let(:attrs) { {status: {"privacyStatus"=>"public"}} }
